--- client/src/pages/Login/LoginCard.jsx
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-const LoginCard = styled.form`
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-export default LoginCard;
